+++ src/main/java/Ava/exceptions/AvaException.java
@@ -1,4 +1,4 @@
-package cloneAva.exceptions;
+package Ava.exceptions;
