+++ src/main/java/Ava/exceptions/CommandNotFoundException.java
@@ -1,4 +1,4 @@
-package cloneAva.exceptions;
+package Ava.exceptions;
 
 
 public class CommandNotFoundException extends AvaException {
