+++ src/main/java/Ava/exceptions/DateTimeNotParsed.java
@@ -1,4 +1,4 @@
-package cloneAva.exceptions;
+package Ava.exceptions;
 
 
 public class DateTimeNotParsed extends AvaException {
