+++ src/main/java/Ava/tasks/Task.java
@@ -1,4 +1,4 @@
-package cloneAva.tasks;
+package Ava.tasks;
 
 
 public class Task {
